+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Change Log
 
+### 4.2.0 (2019/05/02)
+- replace `request` module with `got`
+- changed proxy cli options
+- changed `login` option name to `auth`
+- changed `hls-download` parallel download configuration from 5 parts to 10
+- update modules
+
 ### 4.1.0 (2019/04/05)
 - CLI options for login moved to CUI
 - Removed showing set token at startup